Fix/auto path matching (#13)
* route fix: at startup you go directly to admin login

* feat(auth): implement JWT authentication with cookie support

- Added @nestjs/passport and passport dependencies for JWT authentication.
- Created JwtAuthGuard to protect routes.
- Implemented AuthGuard to validate JWT from cookies.
- Updated AuthService to issue JWTs with a 30-minute expiration.
- Modified AuthController to handle login and logout with cookies.
- Protected admin routes (accommodations, addresses, airports, flights, locations, offers) with JwtAuthGuard.
- Integrated cookie-parser middleware in main.ts for cookie handling.
- Updated module imports to include AuthModule where necessary.

* fix(api): add withCredentials option to HTTP requests for improved security

* interceptor added instead of the prev commit

* fix(api): update offer endpoints to use admin path and add fee handling in entity form

+++ admin-smarttravel-frontend/src/app/services/api.service.ts
@@ -195,14 +195,14 @@ export class ApiService {
   }
 
   createOffer(data: any): Observable<any> {
-    return this._httpClient.post(`http://localhost:3000/api/offers`, data);
+    return this._httpClient.post(`http://localhost:3000/api/admin/offers`, data);
   }
 
   updateOffer(id: string, data: any): Observable<any> {
-    return this._httpClient.patch(`http://localhost:3000/api/offers/${id}`, data);
+    return this._httpClient.patch(`http://localhost:3000/api/admin/offers/${id}`, data);
   }
 
   deleteOffer(id: string): Observable<any> {
-    return this._httpClient.delete(`http://localhost:3000/api/offers/${id}`);
+    return this._httpClient.delete(`http://localhost:3000/api/admin/offers/${id}`);
   }
 }
